주석 추가

+++ src/stakcque/PrintMyJob.java
@@ -15,6 +15,7 @@ public class PrintMyJob {
 
     public static int solution(int[] priorities, int location) {
         int answer = 0;
+        // 필요한 변수 생성 (가장 높은 우선순위, 정렬된 우선순위, 우선순위 리스트)
         Integer biggestNum = 0;
         Queue<Integer> queue = new LinkedList<>();
         ArrayList<Integer> sortedPriorities = new ArrayList();
@@ -25,17 +26,19 @@ public class PrintMyJob {
         sortedPriorities.sort(Collections.reverseOrder());
 
         while(true){
+            // 큐에서 하나씩 빼거나, 빼서 맨 뒤로 보내거나
             biggestNum = sortedPriorities.get(answer);
             if (location == 0 && queue.peek() == biggestNum){
                 answer++;
                 break;
-            } else if ((Integer)queue.peek() < biggestNum){
+            } else if (queue.peek() < biggestNum){
                 queue.add(queue.poll());
             } else {
                 queue.poll();
                 answer++;
             }
 
+            // 내가 선택한 작업의 위치 관리하기
             if(location==0)
                 location = queue.size()-1;
             else
